spannerlib/wrappers/spannerlib-python/spannermockserver/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spannermockserver"
dynamic = ["version"]
authors = [
    { name="Google LLC", email="googleapis-packages@google.com" },
]
description = "A lightweight in-memory mock server for Google Cloud Spanner. This is an internal library that can make breaking changes without prior notice."
readme = "README.md"
license = "Apache-2.0"
license-files = [
    "LICENSE",
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "grpcio>=1.67.0",
    "google-cloud-spanner>=3.55.0", 
    "protobuf>=4.0.0",
    "google-api-core",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "nox",
]

[tool.setuptools]
dynamic = {"version" = {attr = "spannermockserver.__version__"}}

[tool.setuptools.packages.find]
where = ["."]
include = ["spannermockserver*"]

[project.urls]
Homepage = "https://github.com/googleapis/go-sql-spanner/tree/main/spannerlib/wrappers/spannerlib-python/spannermockserver/README.md"
Repository = "https://github.com/googleapis/go-sql-spanner/tree/main/spannerlib/wrappers/spannerlib-python/spannermockserver"
